pyproject.toml
[project]
name = "steel-price"
version = "0.1.0"
description = "Steel spot price data experiments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "akshare>=1.18.46",
    "drissionpage>=4.1.1.2",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "selenium>=4.41.0",
    "tushare>=1.4.29",
]

[tool.uv]
package = false
